Feature/202 auto completion support (#237)
* fix: validate temperature value

* #165 fix output parameter name extraction

* #160 validate project name correctly

* #143 fix crash after removing temperature field (required field)

* #136 use correct abreviation

* #126 fix test selection reordering & update test program UI design

* Enhanc: add test program 'openfile' option & cleanup code generation

* fix missbehaviour when setting bin number

* #116 reorder tests by name

* fix crash when instrument config type doesn't match

* #170 finalize importing labml toolbar (plugin)

* #170 Spyder doesn't filter projects

* semi-ate plugin shall not be triggered for each project type

* #143 clean up handling errors

* #170 fix some design issues

* fix alignment, component size, etc..

* #170 fix some typos & clean up

* #170 fix unit tests

* #132 stop supporting case sensitve naming

* #172 add master app developer support

* user master app parallel with spyder environment
* small refactoring

* #172 update readme file

* #172 fix unittests

* #172 use default device-id in develop mode & handle unexpected test app state

* #171 fix bin table vaidation

* #113 fix plugin config dialog

* #111 fix toolbar target combobox update while startup

* #139 remove not needed flag calculation

* #115 run configuration for all instrument (with and without configuration file)

* use python friendly syntax for naming parameter ('-' character is reserved in python)

* #139 resolve review comments

* #171 improve bin table validation handling

* #172 add missing state handling

* #171 clean up

* #188 fix reseting measuremnt when mpr flag is set

* suspend resolver feature as it do not support mpr feature

* #188 fix unittests

* #188 resolve review comments & fix write measurement issue

* #190 fix device wizard crash

* #172 fix measurement evaluation

* clean up

* update db structure

* #193 add migration script

* #193 add assert info debug

* clean up

* #193 update hardware list

* #161 extend test hook spec

* #161 update/clean up flatcache interface

* #173 single autoscript for a hw configuration

* extend tester interface with setup and teardown functionality

* #202 add autocompletion support

* #202 add project migrration

* #202 update template

* no need to install the test progrram as a package (for now) as we are able to
to include all we need with relative paths

* #226 fix update tree view after target generation & update semi-ate version

* #226 fix update tree view after target generation & update semi-ate v… (#232)

* #226 fix update tree view after target generation & update semi-ate version

* #208 add missing calls to propagate stdf data

* Enh/230 dual die (#233)

* #226 fix update tree view after target generation & update semi-ate version

* #230 fix dual die activation & enable pins tab


* Enh/231 group combo (#234)

* #226 fix update tree view after target generation & update semi-ate version

* #231 delete group shall update the combo box

* PAF-231 fix group dropdown issues

* add more space between items
* adjust width dynamically


* #202 migrate version to fix test export

* #202 fix version structure

* #202 update db version

* [NO-ISSUE] add missing db token for version

* #202 update db version

+++ src/ATE_sammy/ate_sammy/migration/migration_scripts/migration_version_8.py
@@ -10,7 +10,7 @@ class MigrationVersion8(MigratorBase):
             Path(defs_path).joinpath('hardware'), Path(defs_path).joinpath('masksets'), Path(defs_path).joinpath('die'),
             Path(defs_path).joinpath('package'), Path(defs_path).joinpath('device'), Path(defs_path).joinpath('product'),
             Path(defs_path).joinpath('group'), Path(defs_path).joinpath('program'), Path(defs_path).joinpath('qualification'),
-            Path(defs_path).joinpath('test')
+            Path(defs_path).joinpath('test'), Path(defs_path).joinpath('testtarget')
         ]
 
         for category_config_path in all_categories_config_path:
@@ -71,5 +71,9 @@ class MigrationVersion8(MigratorBase):
         for key, values in target_info.items():
             self.write_configuration(key, values)
 
-        import os
-        os.remove(file_name)
+        file_base_name = Path(file_name).stem
+        parent_folder_name = Path(file_name).parent.name
+
+        if (file_base_name == parent_folder_name):
+            import os
+            os.remove(file_name)
